python-ai/requirements.txt
pdfplumber
pytesseract
pdf2image
groq
opencv-python
numpy
Pillow
fastapi
uvicorn
pydantic
python-dotenv
